+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/PluginUtil.kt
@@ -3,8 +3,10 @@ package org.jlleitschuh.gradle.ktlint
 import net.swiftzer.semver.SemVer
 import org.gradle.api.Project
 import org.gradle.api.Task
+import org.gradle.api.file.FileCollection
 import org.gradle.api.plugins.HelpTasksPlugin
 import org.gradle.language.base.plugins.LifecycleBasePlugin
+import java.nio.file.Path
 
 internal fun createConfiguration(target: Project, extension: KtlintExtension) =
     target.configurations.maybeCreate("ktlint").apply {
@@ -20,6 +22,56 @@ internal fun createConfiguration(target: Project, extension: KtlintExtension) =
 
 internal inline fun <reified T : Task> Project.taskHelper(name: String, noinline configuration: T.() -> Unit): T {
     return this.tasks.create(name, T::class.java, configuration)
+}
+
+internal const val EDITOR_CONFIG_FILE_NAME = ".editorconfig"
+
+internal fun getEditorConfigFiles(currentProject: Project): FileCollection {
+    return searchEditorConfigFiles(
+        currentProject,
+        currentProject.projectDir.toPath(),
+        currentProject.files()
+    )
+}
+
+private tailrec fun searchEditorConfigFiles(
+    project: Project,
+    projectPath: Path,
+    fileCollection: FileCollection
+): FileCollection {
+    val editorConfigFC = projectPath.resolve(EDITOR_CONFIG_FILE_NAME)
+    val outputCollection = if (editorConfigFC != null) {
+        fileCollection.plus(project.files(editorConfigFC))
+    } else {
+        fileCollection
+    }
+
+    val parentDir = projectPath.parent
+    return if (parentDir != null &&
+        parentDir != project.rootDir.toPath() &&
+        !editorConfigFC.isRootEditorConfig()) {
+        searchEditorConfigFiles(project, parentDir, outputCollection)
+    } else {
+        outputCollection
+    }
+}
+
+private val editorConfigRootRegex = "^root\\s?=\\s?true\\n".toRegex()
+
+private fun Path.isRootEditorConfig(): Boolean {
+    val asFile = toFile()
+    if (!asFile.exists() || !asFile.canRead()) return false
+
+    val reader = asFile.bufferedReader()
+    var fileLine = reader.readLine()
+    while (fileLine != null) {
+        if (fileLine.contains(editorConfigRootRegex)) {
+            return true
+        }
+        fileLine = reader.readLine()
+    }
+
+    return false
 }
 
 /**